Refactor onclick for notify
